refactor(approval-gate): move pub(crate)-using tests inline
First step of relocating the 2580-line inline test block from
lib.rs into more focused homes. This commit moves the 18 tests
that touch pub(crate) helpers — ones that can't live in tests/*.rs
because integration tests can only see pub items — into the
#[cfg(test)] mod tests block of whichever production module owns
the helper they exercise.

intercept.rs gained:
- interpret_classifier_reply_reads_decision_tags
- decide_intercept_action_* (5 tests)
- apply_policy_rules_* (5 tests)

state.rs gained:
- merge_from_approval_* (4 tests)
- rule_for_returns_matching_rule / rule_for_returns_none_when_absent

sweeper.rs gained:
- timeout_resolved_event_shape

lib.rs loses those 18 tests. 141 approval-gate tests pass (no
duplicate counts — each test runs once). turn-orchestrator + harness
+ shell verified green.

Public-API tests (the remaining ~96 in lib.rs) move to tests/*.rs
in a follow-up commit so they can share fakes via a common module.
