+++ approval-gate/src/lib.rs
@@ -676,236 +676,6 @@ mod tests {
     }
 
     #[test]
-    fn interpret_classifier_reply_reads_decision_tags() {
-        assert!(matches!(
-            interpret_classifier_reply(&json!({"decision": "auto"}), "shell::classify_argv"),
-            Ok(ClassifierDecision::Auto)
-        ));
-        match interpret_classifier_reply(
-            &json!({"decision":"deny","reason":"nope"}),
-            "shell::classify_argv",
-        ) {
-            Ok(ClassifierDecision::Deny(Denial::Policy {
-                classifier_reason,
-                classifier_fn,
-            })) => {
-                assert_eq!(classifier_reason, "nope");
-                assert_eq!(classifier_fn, "shell::classify_argv");
-            }
-            o => panic!("expected Policy denial {:?}", o),
-        }
-        assert!(matches!(
-            interpret_classifier_reply(
-                &json!({"decision":"ask","summary":"x"}),
-                "shell::classify_argv"
-            ),
-            Ok(ClassifierDecision::Ask)
-        ));
-        assert!(interpret_classifier_reply(&json!({}), "shell::classify_argv").is_err());
-    }
-
-    #[test]
-    fn merge_from_approval_inserts_marker_when_inject_true() {
-        let m = merge_from_approval_marker_if_needed(
-            true,
-            json!({"command": "git"}),
-            "call-1",
-            "sess-1",
-        );
-        let inner = m.get("__from_approval").unwrap();
-        assert_eq!(inner["call_id"], "call-1");
-        assert_eq!(inner["session_id"], "sess-1");
-        assert_eq!(m["command"], "git");
-    }
-
-    #[test]
-    fn merge_from_approval_noop_when_inject_false() {
-        let j = json!({"a": 1});
-        let out = merge_from_approval_marker_if_needed(false, j.clone(), "c", "s");
-        assert_eq!(out, j);
-    }
-
-    #[test]
-    fn rule_for_returns_matching_rule() {
-        let rules = vec![
-            InterceptorRule {
-                function_id: "shell::exec".into(),
-                classifier: Some("shell::classify_argv".into()),
-                classifier_timeout_ms: 2000,
-                inject_approval_marker: true,
-                marker_target_verified: true,
-            },
-            InterceptorRule {
-                function_id: "other::fn".into(),
-                classifier: None,
-                classifier_timeout_ms: 2000,
-                inject_approval_marker: false,
-                marker_target_verified: false,
-            },
-        ];
-        let r = rule_for(&rules, "shell::exec").expect("match");
-        assert_eq!(r.classifier.as_deref(), Some("shell::classify_argv"));
-        assert!(r.inject_approval_marker);
-    }
-
-    #[test]
-    fn rule_for_returns_none_when_absent() {
-        let rules = vec![InterceptorRule {
-            function_id: "x::y".into(),
-            classifier: None,
-            classifier_timeout_ms: 2000,
-            inject_approval_marker: false,
-            marker_target_verified: false,
-        }];
-        assert!(rule_for(&rules, "missing::id").is_none());
-    }
-
-    /// An operator-registered rule is authoritative: every call to that
-    /// function id runs through the classifier, even when the run's
-    /// `approval_required` list is empty. This is the inverted contract
-    /// vs. the original "approval_required ANDs the rule" gate.
-    #[test]
-    fn decide_intercept_action_classifies_when_rule_has_classifier_regardless_of_approval_required() {
-        let rule = InterceptorRule {
-            function_id: "shell::exec".into(),
-            classifier: Some("shell::classify_argv".into()),
-            classifier_timeout_ms: 2000,
-            inject_approval_marker: true,
-            marker_target_verified: true,
-        };
-        let action = decide_intercept_action(Some(&rule), false);
-        assert_eq!(
-            action,
-            InterceptAction::Classify {
-                classifier_fn: "shell::classify_argv".into(),
-                classifier_timeout_ms: 2000,
-            }
-        );
-        assert_eq!(action, decide_intercept_action(Some(&rule), true));
-    }
-
-    #[test]
-    fn decide_intercept_action_pauses_when_rule_has_no_classifier_regardless_of_approval_required() {
-        let rule = InterceptorRule {
-            function_id: "shell::fs::write".into(),
-            classifier: None,
-            classifier_timeout_ms: 2000,
-            inject_approval_marker: false,
-            marker_target_verified: false,
-        };
-        assert_eq!(
-            decide_intercept_action(Some(&rule), false),
-            InterceptAction::Pause
-        );
-        assert_eq!(
-            decide_intercept_action(Some(&rule), true),
-            InterceptAction::Pause
-        );
-    }
-
-    #[test]
-    fn decide_intercept_action_pauses_when_no_rule_but_run_listed_approval_required() {
-        assert_eq!(decide_intercept_action(None, true), InterceptAction::Pause);
-    }
-
-    #[test]
-    fn decide_intercept_action_passes_when_no_rule_and_not_approval_required() {
-        assert_eq!(decide_intercept_action(None, false), InterceptAction::Pass);
-    }
-
-    #[test]
-    fn apply_policy_rules_empty_ruleset_falls_through() {
-        let rs: rules::Ruleset = vec![];
-        assert_eq!(
-            apply_policy_rules(&rs, "shell::exec"),
-            PolicyOutcome::FallThrough
-        );
-    }
-
-    #[test]
-    fn apply_policy_rules_allow_short_circuits() {
-        let rs: rules::Ruleset = vec![rules::Rule {
-            permission: "shell::exec".into(),
-            pattern: "*".into(),
-            action: rules::Action::Allow,
-        }];
-        assert_eq!(
-            apply_policy_rules(&rs, "shell::exec"),
-            PolicyOutcome::Allow
-        );
-    }
-
-    #[test]
-    fn apply_policy_rules_deny_carries_matched_rule_identity() {
-        let rs: rules::Ruleset = vec![rules::Rule {
-            permission: "shell::*".into(),
-            pattern: "*".into(),
-            action: rules::Action::Deny,
-        }];
-        assert_eq!(
-            apply_policy_rules(&rs, "shell::fs::write"),
-            PolicyOutcome::Deny {
-                rule_permission: "shell::*".into(),
-                rule_pattern: "*".into(),
-            }
-        );
-    }
-
-    #[test]
-    fn apply_policy_rules_ask_falls_through_to_interceptor_flow() {
-        // Ask means "no decision from this layer — let the next handle it".
-        let rs: rules::Ruleset = vec![rules::Rule {
-            permission: "shell::exec".into(),
-            pattern: "*".into(),
-            action: rules::Action::Ask,
-        }];
-        assert_eq!(
-            apply_policy_rules(&rs, "shell::exec"),
-            PolicyOutcome::FallThrough
-        );
-    }
-
-    #[test]
-    fn apply_policy_rules_last_matching_wins() {
-        // Later-listed more-specific rule overrides earlier permissive default.
-        let rs: rules::Ruleset = vec![
-            rules::Rule {
-                permission: "*".into(),
-                pattern: "*".into(),
-                action: rules::Action::Allow,
-            },
-            rules::Rule {
-                permission: "shell::exec".into(),
-                pattern: "*".into(),
-                action: rules::Action::Deny,
-            },
-        ];
-        assert!(matches!(
-            apply_policy_rules(&rs, "shell::exec"),
-            PolicyOutcome::Deny { .. }
-        ));
-        assert_eq!(
-            apply_policy_rules(&rs, "approval::resolve"),
-            PolicyOutcome::Allow
-        );
-    }
-
-    #[test]
-    fn decide_intercept_action_classifier_empty_string_treated_as_no_classifier() {
-        let rule = InterceptorRule {
-            function_id: "shell::exec".into(),
-            classifier: Some(String::new()),
-            classifier_timeout_ms: 2000,
-            inject_approval_marker: false,
-            marker_target_verified: false,
-        };
-        assert_eq!(
-            decide_intercept_action(Some(&rule), false),
-            InterceptAction::Pause
-        );
-    }
-
-    #[test]
     fn is_terminal_status_returns_true_for_terminal_states() {
         assert!(is_terminal_status("executed"));
         assert!(is_terminal_status("failed"));
@@ -2060,19 +1830,6 @@ mod tests {
     }
 
     #[test]
-    fn timeout_resolved_event_shape() {
-        let evt = timeout_resolved_event("tc-1");
-        assert_eq!(evt["type"], "approval_resolved");
-        assert_eq!(evt["function_call_id"], "tc-1");
-        assert_eq!(evt["tool_call_id"], "tc-1");
-        assert_eq!(evt["decision"], "deny");
-        assert_eq!(evt["status"], "timed_out");
-        // timed_out is self-describing — no Denial / no legacy reason.
-        assert!(evt.get("decision_reason").is_none());
-        assert!(evt.get("denial").is_none());
-    }
-
-    #[test]
     fn unverified_marker_targets_lists_unasserted_rules() {
         let rules = vec![
             InterceptorRule {
@@ -2125,29 +1882,6 @@ mod tests {
     //
     // Each test corresponds to a mutant the test suite previously didn't
     // catch. Test name → mutated line in src/lib.rs.
-
-    #[test]
-    fn merge_from_approval_wraps_null_args_in_marker_only() {
-        // mutant L48: replace `other.is_null()` match guard
-        let out = merge_from_approval_marker_if_needed(true, Value::Null, "c1", "s1");
-        assert!(out.get("__from_approval").is_some());
-        assert!(
-            out.get("payload").is_none(),
-            "null-arg branch must NOT wrap as payload"
-        );
-    }
-
-    #[test]
-    fn merge_from_approval_wraps_scalar_args_in_payload() {
-        // mutant L48: same guard, the other branch
-        let out = merge_from_approval_marker_if_needed(true, json!("scalar"), "c1", "s1");
-        assert!(out.get("__from_approval").is_some());
-        assert_eq!(
-            out.get("payload"),
-            Some(&json!("scalar")),
-            "scalar-arg branch must wrap original under `payload`"
-        );
-    }
 
     #[tokio::test]
     async fn handle_intercept_replay_of_terminal_record_returns_already_resolved() {
