+++ approval-gate/src/state.rs
@@ -177,3 +177,81 @@ pub fn unverified_marker_targets(rules: &[InterceptorRule]) -> Vec<&str> {
         .map(|r| r.function_id.as_str())
         .collect()
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use serde_json::json;
+
+    #[test]
+    fn merge_from_approval_inserts_marker_when_inject_true() {
+        let m = merge_from_approval_marker_if_needed(
+            true,
+            json!({"command": "git"}),
+            "call-1",
+            "sess-1",
+        );
+        let inner = m.get("__from_approval").unwrap();
+        assert_eq!(inner["call_id"], "call-1");
+        assert_eq!(inner["session_id"], "sess-1");
+        assert_eq!(m["command"], "git");
+    }
+
+    #[test]
+    fn merge_from_approval_noop_when_inject_false() {
+        let j = json!({"a": 1});
+        let out = merge_from_approval_marker_if_needed(false, j.clone(), "c", "s");
+        assert_eq!(out, j);
+    }
+
+    #[test]
+    fn merge_from_approval_wraps_null_args_in_marker_only() {
+        let m = merge_from_approval_marker_if_needed(true, Value::Null, "c1", "s1");
+        let obj = m.as_object().unwrap();
+        assert_eq!(obj.len(), 1);
+        assert!(obj.contains_key("__from_approval"));
+    }
+
+    #[test]
+    fn merge_from_approval_wraps_scalar_args_in_payload() {
+        let out = merge_from_approval_marker_if_needed(true, json!("scalar"), "c1", "s1");
+        assert_eq!(out["payload"], json!("scalar"));
+        assert_eq!(out["__from_approval"]["call_id"], "c1");
+        assert_eq!(out["__from_approval"]["session_id"], "s1");
+    }
+
+    #[test]
+    fn rule_for_returns_matching_rule() {
+        let rules = vec![
+            InterceptorRule {
+                function_id: "shell::exec".into(),
+                classifier: Some("shell::classify_argv".into()),
+                classifier_timeout_ms: 2000,
+                inject_approval_marker: true,
+                marker_target_verified: true,
+            },
+            InterceptorRule {
+                function_id: "other::fn".into(),
+                classifier: None,
+                classifier_timeout_ms: 2000,
+                inject_approval_marker: false,
+                marker_target_verified: false,
+            },
+        ];
+        let r = rule_for(&rules, "shell::exec").expect("match");
+        assert_eq!(r.classifier.as_deref(), Some("shell::classify_argv"));
+        assert!(r.inject_approval_marker);
+    }
+
+    #[test]
+    fn rule_for_returns_none_when_absent() {
+        let rules = vec![InterceptorRule {
+            function_id: "x::y".into(),
+            classifier: None,
+            classifier_timeout_ms: 2000,
+            inject_approval_marker: false,
+            marker_target_verified: false,
+        }];
+        assert!(rule_for(&rules, "missing::id").is_none());
+    }
+}
